[feat] #22 사진 촬영 뷰 구현
